backend/requirements.txt
Flask==3.1.0
Flask_Cors==4.0.1
Flask_JWT_Extended==4.7.1
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.36
bcrypt==4.2.1
pytest==7.4.4
python-dotenv==1.1.0